Add Board interfaces and documentation
This commit adds additional board interfaces which
describe different ways that a board may render.
This allows for flexibility and provides consistent
ways for the core app to render the boards provided
by the plugins, while ensuring that no additional
3rd-party libraries are needed by plugins.

Additionally, some inline documentation has been
added to the new board interfaces as well as a
couple of other places in the library.